Separate out schedule_utils.{cc,h} from schedule_transforms.{cc,h}
Many of the functions in schedule_transforms.{cc,h} are not
transformations but some higher-level functions that operate on (parts
of) schedule trees rather than on individual nodes.  These functions
were introduced into schedule_transforms.{cc,h} for historical reasons.
The file has grown large.  Separate the non-transformation functions
into separate files, schedule_utils.{cc,h}.
