+++ tc/core/CMakeLists.txt
@@ -29,6 +29,7 @@ add_library(
   polyhedral/schedule_tree.cc
   polyhedral/schedule_tree_elem.cc
   polyhedral/schedule_print.cc
+  polyhedral/schedule_utils.cc
   polyhedral/scop.cc
   polyhedral/separation.cc
   polyhedral/unroll.cc
