+++ tc/core/polyhedral/cuda/mapped_scop.cc
@@ -35,6 +35,7 @@
 #include "tc/core/polyhedral/functional.h"
 #include "tc/core/polyhedral/schedule_transforms.h"
 #include "tc/core/polyhedral/schedule_tree_matcher.h"
+#include "tc/core/polyhedral/schedule_utils.h"
 #include "tc/core/polyhedral/scop.h"
 #include "tc/core/polyhedral/separation.h"
 #include "tc/core/polyhedral/unroll.h"
